main: implement ONLY UI for jmp
When building the program table, it stores label name in the table.
And program meets the jmp instruction, it finds the label name
and the line number of the label. And then it responses the next
line-number of the label to the web page. Then the web-page sets
the highlight at the next line of the label.

This patch implements only the UI part of the jmp instruction.